Move Git editor manager to separate translation unit
